Moved character selection into the room/id page.
