+++ package.json
@@ -59,6 +59,7 @@
     "react-resizable-panels": "^3.0.1",
     "react-router-dom": "^7.5.3",
     "recharts": "^2.15.3",
+    "sockjs": "^0.3.24",
     "sockjs-client": "^1.6.1",
     "sonner": "^2.0.3",
     "tailwind-merge": "^3.2.0",
