+++ src/routeTree.gen.ts
@@ -21,16 +21,12 @@ import { Route as publicPublicLoginImport } from './routes/(public)/_public.logi
 import { Route as authAuthRulesImport } from './routes/(auth)/_auth.rules'
 import { Route as authAuthLobbyImport } from './routes/(auth)/_auth.lobby'
 import { Route as authAuthCharacterPageImport } from './routes/(auth)/_auth.character-page'
-import { Route as authlobbyLobbyImport } from './routes/(auth)/(lobby)/_lobby'
 import { Route as authAuthRoomRoomIdImport } from './routes/(auth)/_auth.room.$roomId'
-import { Route as authlobbyLobbyCharacterSelectImport } from './routes/(auth)/(lobby)/_lobby.character-select'
-import { Route as authlobbyLobbyWaitingRoomLobbyIdImport } from './routes/(auth)/(lobby)/_lobby.waiting-room.$lobbyId'
 
 // Create Virtual Routes
 
 const publicImport = createFileRoute('/(public)')()
 const authImport = createFileRoute('/(auth)')()
-const authlobbyImport = createFileRoute('/(auth)/(lobby)')()
 
 // Create/Update Routes
 
@@ -42,11 +38,6 @@ const publicRoute = publicImport.update({
 const authRoute = authImport.update({
   id: '/(auth)',
   getParentRoute: () => rootRoute,
-} as any)
-
-const authlobbyRoute = authlobbyImport.update({
-  id: '/(lobby)',
-  getParentRoute: () => authRoute,
 } as any)
 
 const publicPublicRoute = publicPublicImport.update({
@@ -95,30 +86,11 @@ const authAuthCharacterPageRoute = authAuthCharacterPageImport.update({
   getParentRoute: () => authAuthRoute,
 } as any)
 
-const authlobbyLobbyRoute = authlobbyLobbyImport.update({
-  id: '/_lobby',
-  getParentRoute: () => authlobbyRoute,
-} as any)
-
 const authAuthRoomRoomIdRoute = authAuthRoomRoomIdImport.update({
   id: '/room/$roomId',
   path: '/room/$roomId',
   getParentRoute: () => authAuthRoute,
 } as any)
-
-const authlobbyLobbyCharacterSelectRoute =
-  authlobbyLobbyCharacterSelectImport.update({
-    id: '/character-select',
-    path: '/character-select',
-    getParentRoute: () => authlobbyLobbyRoute,
-  } as any)
-
-const authlobbyLobbyWaitingRoomLobbyIdRoute =
-  authlobbyLobbyWaitingRoomLobbyIdImport.update({
-    id: '/waiting-room/$lobbyId',
-    path: '/waiting-room/$lobbyId',
-    getParentRoute: () => authlobbyLobbyRoute,
-  } as any)
 
 // Populate the FileRoutesByPath interface
 
@@ -151,20 +123,6 @@ declare module '@tanstack/react-router' {
       fullPath: '/'
       preLoaderRoute: typeof publicPublicImport
       parentRoute: typeof publicRoute
-    }
-    '/(auth)/(lobby)': {
-      id: '/(auth)/(lobby)'
-      path: '/'
-      fullPath: '/'
-      preLoaderRoute: typeof authlobbyImport
-      parentRoute: typeof authImport
-    }
-    '/(auth)/(lobby)/_lobby': {
-      id: '/(auth)/(lobby)/_lobby'
-      path: '/'
-      fullPath: '/'
-      preLoaderRoute: typeof authlobbyLobbyImport
-      parentRoute: typeof authlobbyRoute
     }
     '/(auth)/_auth/character-page': {
       id: '/(auth)/_auth/character-page'
@@ -208,26 +166,12 @@ declare module '@tanstack/react-router' {
       preLoaderRoute: typeof authAuthIndexImport
       parentRoute: typeof authAuthImport
     }
-    '/(auth)/(lobby)/_lobby/character-select': {
-      id: '/(auth)/(lobby)/_lobby/character-select'
-      path: '/character-select'
-      fullPath: '/character-select'
-      preLoaderRoute: typeof authlobbyLobbyCharacterSelectImport
-      parentRoute: typeof authlobbyLobbyImport
-    }
     '/(auth)/_auth/room/$roomId': {
       id: '/(auth)/_auth/room/$roomId'
       path: '/room/$roomId'
       fullPath: '/room/$roomId'
       preLoaderRoute: typeof authAuthRoomRoomIdImport
       parentRoute: typeof authAuthImport
-    }
-    '/(auth)/(lobby)/_lobby/waiting-room/$lobbyId': {
-      id: '/(auth)/(lobby)/_lobby/waiting-room/$lobbyId'
-      path: '/waiting-room/$lobbyId'
-      fullPath: '/waiting-room/$lobbyId'
-      preLoaderRoute: typeof authlobbyLobbyWaitingRoomLobbyIdImport
-      parentRoute: typeof authlobbyLobbyImport
     }
   }
 }
@@ -254,40 +198,12 @@ const authAuthRouteWithChildren = authAuthRoute._addFileChildren(
   authAuthRouteChildren,
 )
 
-interface authlobbyLobbyRouteChildren {
-  authlobbyLobbyCharacterSelectRoute: typeof authlobbyLobbyCharacterSelectRoute
-  authlobbyLobbyWaitingRoomLobbyIdRoute: typeof authlobbyLobbyWaitingRoomLobbyIdRoute
-}
-
-const authlobbyLobbyRouteChildren: authlobbyLobbyRouteChildren = {
-  authlobbyLobbyCharacterSelectRoute: authlobbyLobbyCharacterSelectRoute,
-  authlobbyLobbyWaitingRoomLobbyIdRoute: authlobbyLobbyWaitingRoomLobbyIdRoute,
-}
-
-const authlobbyLobbyRouteWithChildren = authlobbyLobbyRoute._addFileChildren(
-  authlobbyLobbyRouteChildren,
-)
-
-interface authlobbyRouteChildren {
-  authlobbyLobbyRoute: typeof authlobbyLobbyRouteWithChildren
-}
-
-const authlobbyRouteChildren: authlobbyRouteChildren = {
-  authlobbyLobbyRoute: authlobbyLobbyRouteWithChildren,
-}
-
-const authlobbyRouteWithChildren = authlobbyRoute._addFileChildren(
-  authlobbyRouteChildren,
-)
-
 interface authRouteChildren {
   authAuthRoute: typeof authAuthRouteWithChildren
-  authlobbyRoute: typeof authlobbyRouteWithChildren
 }
 
 const authRouteChildren: authRouteChildren = {
   authAuthRoute: authAuthRouteWithChildren,
-  authlobbyRoute: authlobbyRouteWithChildren,
 }
 
 const authRouteWithChildren = authRoute._addFileChildren(authRouteChildren)
@@ -324,9 +240,7 @@ export interface FileRoutesByFullPath {
   '/rules': typeof authAuthRulesRoute
   '/login': typeof publicPublicLoginRoute
   '/register': typeof publicPublicRegisterRoute
-  '/character-select': typeof authlobbyLobbyCharacterSelectRoute
   '/room/$roomId': typeof authAuthRoomRoomIdRoute
-  '/waiting-room/$lobbyId': typeof authlobbyLobbyWaitingRoomLobbyIdRoute
 }
 
 export interface FileRoutesByTo {
@@ -336,9 +250,7 @@ export interface FileRoutesByTo {
   '/rules': typeof authAuthRulesRoute
   '/login': typeof publicPublicLoginRoute
   '/register': typeof publicPublicRegisterRoute
-  '/character-select': typeof authlobbyLobbyCharacterSelectRoute
   '/room/$roomId': typeof authAuthRoomRoomIdRoute
-  '/waiting-room/$lobbyId': typeof authlobbyLobbyWaitingRoomLobbyIdRoute
 }
 
 export interface FileRoutesById {
@@ -347,17 +259,13 @@ export interface FileRoutesById {
   '/(auth)/_auth': typeof authAuthRouteWithChildren
   '/(public)': typeof publicRouteWithChildren
   '/(public)/_public': typeof publicPublicRouteWithChildren
-  '/(auth)/(lobby)': typeof authlobbyRouteWithChildren
-  '/(auth)/(lobby)/_lobby': typeof authlobbyLobbyRouteWithChildren
   '/(auth)/_auth/character-page': typeof authAuthCharacterPageRoute
   '/(auth)/_auth/lobby': typeof authAuthLobbyRoute
   '/(auth)/_auth/rules': typeof authAuthRulesRoute
   '/(public)/_public/login': typeof publicPublicLoginRoute
   '/(public)/_public/register': typeof publicPublicRegisterRoute
   '/(auth)/_auth/': typeof authAuthIndexRoute
-  '/(auth)/(lobby)/_lobby/character-select': typeof authlobbyLobbyCharacterSelectRoute
   '/(auth)/_auth/room/$roomId': typeof authAuthRoomRoomIdRoute
-  '/(auth)/(lobby)/_lobby/waiting-room/$lobbyId': typeof authlobbyLobbyWaitingRoomLobbyIdRoute
 }
 
 export interface FileRouteTypes {
@@ -369,9 +277,7 @@ export interface FileRouteTypes {
     | '/rules'
     | '/login'
     | '/register'
-    | '/character-select'
     | '/room/$roomId'
-    | '/waiting-room/$lobbyId'
   fileRoutesByTo: FileRoutesByTo
   to:
     | '/'
@@ -380,26 +286,20 @@ export interface FileRouteTypes {
     | '/rules'
     | '/login'
     | '/register'
-    | '/character-select'
     | '/room/$roomId'
-    | '/waiting-room/$lobbyId'
   id:
     | '__root__'
     | '/(auth)'
     | '/(auth)/_auth'
     | '/(public)'
     | '/(public)/_public'
-    | '/(auth)/(lobby)'
-    | '/(auth)/(lobby)/_lobby'
     | '/(auth)/_auth/character-page'
     | '/(auth)/_auth/lobby'
     | '/(auth)/_auth/rules'
     | '/(public)/_public/login'
     | '/(public)/_public/register'
     | '/(auth)/_auth/'
-    | '/(auth)/(lobby)/_lobby/character-select'
     | '/(auth)/_auth/room/$roomId'
-    | '/(auth)/(lobby)/_lobby/waiting-room/$lobbyId'
   fileRoutesById: FileRoutesById
 }
 
@@ -430,8 +330,7 @@ export const routeTree = rootRoute
     "/(auth)": {
       "filePath": "(auth)",
       "children": [
-        "/(auth)/_auth",
-        "/(auth)/(lobby)"
+        "/(auth)/_auth"
       ]
     },
     "/(auth)/_auth": {
@@ -459,21 +358,6 @@ export const routeTree = rootRoute
         "/(public)/_public/register"
       ]
     },
-    "/(auth)/(lobby)": {
-      "filePath": "(auth)/(lobby)",
-      "parent": "/(auth)",
-      "children": [
-        "/(auth)/(lobby)/_lobby"
-      ]
-    },
-    "/(auth)/(lobby)/_lobby": {
-      "filePath": "(auth)/(lobby)/_lobby.tsx",
-      "parent": "/(auth)/(lobby)",
-      "children": [
-        "/(auth)/(lobby)/_lobby/character-select",
-        "/(auth)/(lobby)/_lobby/waiting-room/$lobbyId"
-      ]
-    },
     "/(auth)/_auth/character-page": {
       "filePath": "(auth)/_auth.character-page.tsx",
       "parent": "/(auth)/_auth"
@@ -498,17 +382,9 @@ export const routeTree = rootRoute
       "filePath": "(auth)/_auth.index.tsx",
       "parent": "/(auth)/_auth"
     },
-    "/(auth)/(lobby)/_lobby/character-select": {
-      "filePath": "(auth)/(lobby)/_lobby.character-select.tsx",
-      "parent": "/(auth)/(lobby)/_lobby"
-    },
     "/(auth)/_auth/room/$roomId": {
       "filePath": "(auth)/_auth.room.$roomId.tsx",
       "parent": "/(auth)/_auth"
-    },
-    "/(auth)/(lobby)/_lobby/waiting-room/$lobbyId": {
-      "filePath": "(auth)/(lobby)/_lobby.waiting-room.$lobbyId.tsx",
-      "parent": "/(auth)/(lobby)/_lobby"
     }
   }
 }
